Clean code and include a new profile.py for all logic code behind settings customization
